long cast

+++ rate/src/main/java/com/bbou/rate/AppRate.java
@@ -70,7 +70,7 @@ public final class AppRate
 
 	private static boolean isOverDate(long targetDate, int threshold)
 	{
-		return new Date().getTime() - targetDate >= threshold * 24 * 60 * 60 * 1000;
+		return new Date().getTime() - targetDate >= (long) threshold * 24 * 60 * 60 * 1000;
 	}
 
 	@NonNull
